+++ apps/avifgainmaputil/imageio.cc
@@ -190,7 +190,8 @@ avifResult WriteAvifGrid(const avifImage* image, int grid_cols, int grid_rows,
 
 avifResult ReadImage(avifImage* image, const std::string& input_filename,
                      avifPixelFormat requested_format, uint32_t requested_depth,
-                     bool ignore_profile, bool ignore_gain_map, int jobs) {
+                     bool ignore_profile, bool ignore_alpha,
+                     bool ignore_gain_map, int jobs) {
   avifAppFileFormat input_format = avifGuessFileFormat(input_filename.c_str());
   if (input_format == AVIF_APP_FILE_FORMAT_UNKNOWN) {
     std::cerr << "Cannot determine input format: " << input_filename;
@@ -204,7 +205,8 @@ avifResult ReadImage(avifImage* image, const std::string& input_filename,
       decoder->imageContentToDecode |= AVIF_IMAGE_CONTENT_GAIN_MAP;
     }
     decoder->maxThreads = jobs;
-    avifResult result = ReadAvif(decoder.get(), input_filename, ignore_profile);
+    avifResult result =
+        ReadAvif(decoder.get(), input_filename, ignore_profile, ignore_alpha);
     if (result != AVIF_RESULT_OK) {
       return result;
     }
@@ -238,7 +240,7 @@ avifResult ReadImage(avifImage* image, const std::string& input_filename,
         input_filename.c_str(), AVIF_APP_FILE_FORMAT_UNKNOWN /* guess format */,
         requested_format, static_cast<int>(requested_depth),
         AVIF_CHROMA_DOWNSAMPLING_AUTOMATIC, ignore_profile,
-        /*ignoreExif=*/false, /*ignoreXMP=*/false, /*ignoreAlpha=*/false,
+        /*ignoreExif=*/false, /*ignoreXMP=*/false, ignore_alpha,
         ignore_gain_map, AVIF_DEFAULT_IMAGE_SIZE_LIMIT, image,
         /*outDepth=*/nullptr,
         /*sourceTiming=*/nullptr, /*frameIter=*/nullptr);
@@ -263,7 +265,7 @@ avifResult ReadImage(avifImage* image, const std::string& input_filename,
 }
 
 avifResult ReadAvif(avifDecoder* decoder, const std::string& input_filename,
-                    bool ignore_profile) {
+                    bool ignore_profile, bool ignore_alpha) {
   avifResult result = avifDecoderSetIOFile(decoder, input_filename.c_str());
   if (result != AVIF_RESULT_OK) {
     std::cerr << "Cannot open file for read: " << input_filename << "\n";
@@ -286,6 +288,9 @@ avifResult ReadAvif(avifDecoder* decoder, const std::string& input_filename,
     if (decoder->image->gainMap) {
       avifRWDataFree(&decoder->image->gainMap->altICC);
     }
+  }
+  if (ignore_alpha && decoder->image->alphaPlane) {
+    avifImageFreePlanes(decoder->image, AVIF_PLANES_A);
   }
 
   return AVIF_RESULT_OK;