+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The changes are relative to the previous release, unless the baseline is specifi
 ### Added since 1.4.2
 
 * avifenc: add --ignore-alpha flag to discard alpha channel on encode
+* avifgainmaputil: add --ignore-alpha flag to discard alpha channel
 
 ### Changed since 1.4.2